+++ Pandora/Paradigm/Basis.hs
@@ -14,9 +14,6 @@ import Pandora.Paradigm.Basis.Jack as Exports
 import Pandora.Paradigm.Basis.Product as Exports
 import Pandora.Paradigm.Basis.Constant as Exports
 import Pandora.Paradigm.Basis.Identity as Exports
-import Pandora.Paradigm.Basis.Junction.Kan as Exports
-import Pandora.Paradigm.Basis.Junction.Transformer as Exports
-import Pandora.Paradigm.Basis.Junction.Composition as Exports
 
 import Pandora.Core.Functor (type (~>))
 import Pandora.Core.Morphism ((.), (!))